restructured files and removed redundant lib sub directory
